requirements-gpu.txt
# GPU対応の強化学習環境依存関係
# PyTorch GPU版
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchaudio==2.1.2+cu121

# 強化学習フレームワーク
gymnasium[all]==0.29.1
stable-baselines3==2.1.0
ray[rllib]==2.8.0
torchrl==0.2.0

# 分散学習・最適化
accelerate==0.25.0
deepspeed==0.12.6
horovod==0.28.1

# 監視・プロファイリング
tensorboard==2.15.1
wandb==0.16.1
nvidia-ml-py3==7.352.0
psutil==5.9.6

# データ処理・高速化
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
lark==1.1.5

# 可視化
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# 開発・デバッグ
jupyter==1.0.0
ipython==8.15.0
pytest==7.4.3
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# ROS 2 関連
rclpy
ros2
ros_gz_bridge
ros_gz_sim

# その他
tqdm==4.66.1
rich==13.7.0 